Seongryeol [성렬 리뷰페이지 커밋]

+++ src/main/java/com/project/foodiefoodie/review/domain/Review.java
@@ -18,7 +18,7 @@ public class Review {
     private String businessNo;
     private Date lastUpdated;
     private long likeCnt;
-    private double starRate;
+    private int starRate;
     private String isPrivate;
     private int reportCnt;
